Create simple menu, move OneSudoku to its own file
