+++ app/src/main/res/menu/top_bar_menu.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/menuNewGame"
+        android:title="New Game"
+        app:showAsAction="never" />
+
+
+    <item
+        android:id="@+id/menuCloseApp"
+        android:title="Close App"
+        app:showAsAction="never" />
+
+</menu>
